requirements.txt
PyYAML>=3.10
aiohttp>=2.2.0
async-timeout>=1.2.0
chardet>=3.0.0
-e git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
-e git+https://github.com/strinking/python-duckduckgo@async#egg=duckduckgo
multidict>=3.1.0
websockets>=3.0
yarl>=0.12.0
